backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
anthropic==0.40.0
httpx==0.27.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
alembic==1.13.0
